requirements.txt
pyautogui
pyinstaller
pystray
Pillow
customtkinter
